refactor(schema): rename DSL block directory from components to blocks
The content-shape DSL calls them blocks, so the user-authored directory
should be `blocks/` not `components/`. The astro playground and the CLI
`init` emitter still used `components/`, while the vanilla playground
already used `blocks/`.

Renames the astro playground `src/schema/components/` to
`src/schema/blocks/` (with all importers updated), switches `schema init`
to emit blocks into `blocks/`, and points `schema push` directory
discovery at a sibling `blocks/` directory only (no legacy `components/`
fallback). The MAPI wire `component` resource (REST endpoints, snapshot
cache, `SchemaData.components`) is untouched.

Refs WDX-419
