+++ packages/cli/src/commands/schema/init/index.test.ts
@@ -126,7 +126,7 @@ describe('schema init command', () => {
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE]);
 
     const files = Object.keys(vol.toJSON());
-    const componentFile = files.find(f => f.includes('/components/hero.ts'));
+    const componentFile = files.find(f => f.includes('/blocks/hero.ts'));
     expect(componentFile).toBeDefined();
 
     const content = vol.readFileSync(componentFile!, 'utf-8') as string;
@@ -144,9 +144,9 @@ describe('schema init command', () => {
 
     const files = Object.keys(vol.toJSON());
     // The block is placed inside its slugified group directory; no folder file is written.
-    const componentFile = files.find(f => f.includes('/components/my-layout/hero.ts'));
+    const componentFile = files.find(f => f.includes('/blocks/my-layout/hero.ts'));
     expect(componentFile).toBeDefined();
-    expect(files.some(f => f.includes('/components/folders/'))).toBe(false);
+    expect(files.some(f => f.includes('/blocks/folders/'))).toBe(false);
 
     const content = vol.readFileSync(componentFile!, 'utf-8') as string;
     expect(content).toContain('defineBlock(');
@@ -156,7 +156,7 @@ describe('schema init command', () => {
 
     // schema.ts imports the block from its slugified subdirectory.
     const schemaFile = vol.readFileSync(files.find(f => f.endsWith('/schema.ts'))!, 'utf-8') as string;
-    expect(schemaFile).toContain('./components/my-layout/hero');
+    expect(schemaFile).toContain('./blocks/my-layout/hero');
   });
 
   it('should generate datasource files', async () => {
@@ -199,7 +199,7 @@ describe('schema init command', () => {
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE]);
 
     const files = Object.keys(vol.toJSON());
-    expect(files.some(f => f.includes('.storyblok/schema/components/hero.ts'))).toBe(true);
+    expect(files.some(f => f.includes('.storyblok/schema/blocks/hero.ts'))).toBe(true);
     expect(files.some(f => f.includes('.storyblok/schema/schema.ts'))).toBe(true);
   });
 
@@ -210,7 +210,7 @@ describe('schema init command', () => {
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE, '--out-dir', './custom/output']);
 
     const files = Object.keys(vol.toJSON());
-    expect(files.some(f => f.includes('custom/output/components/hero.ts'))).toBe(true);
+    expect(files.some(f => f.includes('custom/output/blocks/hero.ts'))).toBe(true);
     expect(files.some(f => f.includes('custom/output/schema.ts'))).toBe(true);
   });
 
@@ -256,7 +256,7 @@ describe('schema init command', () => {
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE]);
 
     const files = Object.keys(vol.toJSON());
-    const componentFile = files.find(f => f.includes('/components/hero.ts'));
+    const componentFile = files.find(f => f.includes('/blocks/hero.ts'));
     expect(componentFile).toBeDefined();
 
     const content = vol.readFileSync(componentFile!, 'utf-8') as string;
@@ -272,13 +272,13 @@ describe('schema init command', () => {
 
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE]);
 
-    expect(console.log).toHaveBeenCalledWith('  .storyblok/schema/components/hero.ts');
+    expect(console.log).toHaveBeenCalledWith('  .storyblok/schema/blocks/hero.ts');
     expect(console.log).toHaveBeenCalledWith('  .storyblok/schema/schema.ts');
   });
 
   it('should refuse when the target directory is not empty', async () => {
     vol.fromJSON({
-      '.storyblok/schema/components/existing.ts': '// existing',
+      '.storyblok/schema/blocks/existing.ts': '// existing',
     });
     let unhandledRequest = false;
     server.use(
@@ -307,6 +307,6 @@ describe('schema init command', () => {
     await schemaCommand.parseAsync(['node', 'test', 'init', '--space', DEFAULT_SPACE]);
 
     const files = Object.keys(vol.toJSON());
-    expect(files.some(f => f.includes('.storyblok/schema/components/hero.ts'))).toBe(true);
+    expect(files.some(f => f.includes('.storyblok/schema/blocks/hero.ts'))).toBe(true);
   });
 });
