+++ packages/schema/playground/astro/src/components/storyblok/hero.astro
@@ -1,7 +1,7 @@
 ---
 import type { BlockContent } from '@storyblok/schema';
 import type { Blocks } from '../../schema/schema';
-import type { heroBlock } from '../../schema/components/hero';
+import type { heroBlock } from '../../schema/blocks/hero';
 
 type HeroContent = BlockContent<typeof heroBlock, Blocks>;
