+++ packages/schema/playground/astro/src/components/storyblok/intro.astro
@@ -1,7 +1,7 @@
 ---
 import type { BlockContent } from '@storyblok/schema';
 import type { Blocks } from '../../schema/schema';
-import type { introBlock } from '../../schema/components/intro';
+import type { introBlock } from '../../schema/blocks/intro';
 
 type IntroContent = BlockContent<typeof introBlock, Blocks>;
