+++ packages/schema/playground/astro/src/components/storyblok/media.astro
@@ -2,7 +2,7 @@
 import { marked } from 'marked';
 import type { BlockContent } from '@storyblok/schema';
 import type { Blocks } from '../../schema/schema';
-import type { mediaBlock } from '../../schema/components/media';
+import type { mediaBlock } from '../../schema/blocks/media';
 
 type MediaContent = BlockContent<typeof mediaBlock, Blocks>;
