+++ packages/schema/playground/astro/src/components/storyblok/teaser-list.astro
@@ -1,7 +1,7 @@
 ---
 import type { BlockContent } from '@storyblok/schema';
 import type { Blocks } from '../../schema/schema';
-import type { teaserListBlock } from '../../schema/components/teaser-list';
+import type { teaserListBlock } from '../../schema/blocks/teaser-list';
 import Teaser from './teaser.astro';
 
 type TeaserListContent = BlockContent<typeof teaserListBlock, Blocks>;
