+++ packages/schema/playground/astro/src/components/storyblok/teaser.astro
@@ -1,7 +1,7 @@
 ---
 import type { BlockContent } from '@storyblok/schema';
 import type { Blocks } from '../../schema/schema';
-import type { teaserBlock } from '../../schema/components/teaser';
+import type { teaserBlock } from '../../schema/blocks/teaser';
 
 type TeaserContent = BlockContent<typeof teaserBlock, Blocks>;
