+++ packages/schema/playground/astro/src/schema/blocks/article.ts
