+++ packages/schema/playground/astro/src/schema/blocks/banner.ts
