+++ packages/schema/playground/astro/src/schema/blocks/comparison-table.ts
