+++ packages/schema/playground/astro/src/schema/blocks/faq.ts
