+++ packages/schema/playground/astro/src/schema/blocks/gallery.ts
