+++ packages/schema/playground/astro/src/schema/blocks/hero.ts
