+++ packages/schema/playground/astro/src/schema/blocks/media.ts
