+++ packages/schema/playground/astro/src/schema/blocks/page.ts
