+++ packages/schema/playground/astro/src/schema/blocks/stats.ts
