+++ packages/schema/playground/astro/src/schema/blocks/teaser-list.ts
