+++ packages/schema/playground/astro/src/seeds/stories/about.ts
@@ -1,9 +1,9 @@
 import { createStoryHelpers } from '@storyblok/schema';
 import type { Asset } from '@storyblok/management-api-client';
-import { pageBlock } from '../../schema/components/page';
-import { heroBlock } from '../../schema/components/hero';
-import { introBlock } from '../../schema/components/intro';
-import { mediaBlock } from '../../schema/components/media';
+import { pageBlock } from '../../schema/blocks/page';
+import { heroBlock } from '../../schema/blocks/hero';
+import { introBlock } from '../../schema/blocks/intro';
+import { mediaBlock } from '../../schema/blocks/media';
 import type { Schema } from '../../schema/schema';
 
 const { defineStoryCreate } = createStoryHelpers().withTypes<Schema>();
